[REFACTOR] Use singleton pattern for FlagRegister
Use singleton pattern to keep a single instance of FlagRegister around.

Prevents needing to create a new FlagRegister every time we need to
check a flag or add a flag.

May impact testability of the plugin.

Alternative might be using dependency injection.

Reading this made me think this might be a good use case for the
singleton design pattern.

https://enterprisecraftsmanship.com/posts/singleton-vs-dependency-injection/