[FEATURE] Implement function to check flag status
Adds a simple function to check the status of the flag.

This function will only be available from plugins which are loaded _after_
this plugin.

Using this function for a flag which has not yet been registered will
result in a fatal error. Any calls to feature_flag_enabled will need to
be made after the given flag has been added to the register_flags
action.

I've made this a global function, but I could move it to a static
function, and then we just call something like
FlagsHelper::feature_flag_enabled( $flag_key ) - feedback would be well
appreciated.